Initial commit
This is the result of experimenting and "re-learning" Hugo for like the
fourth time. I think setting up a TIL micro (pico?) blog is a nice way
to start writing more and getting some stuff online.

This initial version took about 2 days: one to get acquainted with
templates and coming up with a cohesive semantic layout and another one
to experiment with stylesheets, fonts, and general design stuff.

There's still a lot that I feel is missing: a navigation menu (currently
would only like back to the homepage), a view of existing tags on the
page (possibly as an aside?), a summary/tagline on the homepage before
the entry list, an about page, proper style editing for code, add
pusblish date and tags to a post's header, and probably many more.
I feel like most of the above are very optional, and the rest are best
left to be worked on as I create more content and can see what the
appropriate direction is.

This feels like a decent start so that I can write short posts in
Markdown and see nifty and fun HTML come out. :)